requirements.txt
Pillow
pynput
pyautogui
pystray